+++ DotDoc.EntityFrameworkCore.Extensions.Tests/DoesExistExtensions/DoesDatabaseExistTests.cs
@@ -19,7 +19,7 @@ public class DoesDatabaseExistTests
     /// </summary>
     /// <param name="databaseType">Database type.</param>
     /// <param name="useAsync">If <see langword="true"/> then tests the async method.</param>
-    /// <returns><see cref="Task"/>.</returns>
+    /// <returns>A task that represents the asynchronous test operation.</returns>
     [TestMethod]
     [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite DoesDatabaseExist when database exists.")]
     [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite DoesDatabaseExistAsync when database exists.")]
@@ -42,7 +42,7 @@ public class DoesDatabaseExistTests
     /// <param name="databaseType">Database type.</param>
     /// <param name="connectionString">Connection string.</param>
     /// <param name="useAsync">If <see langword="true"/> then tests the async method.</param>
-    /// <returns><see cref="Task"/>.</returns>
+    /// <returns>A task that represents the asynchronous test operation.</returns>
     [TestMethod]
     [DataRow(DatabaseType.Sqlite, "Data Source = c:\\NonExistantDatabase.db", false, DisplayName = "SQLite DoesDatabaseExist when database does not exist.")]
     [DataRow(DatabaseType.Sqlite, "Data Source = c:\\NonExistantDatabase.db", true, DisplayName = "SQLite DoesDatabaseExistAsync when database does not exist.")]