+++ DotDoc.EntityFrameworkCore.Extensions.Tests/QueryExtensions/ExecuteScalarTests.cs
@@ -19,48 +19,78 @@ public class ExecuteScalarTests
     /// </summary>
     /// <param name="databaseType">Database type.</param>
     /// <param name="useAsync">If <see langword="true"/> then tests the async method.</param>
-    /// <returns><see cref="Task"/>.</returns>
+    /// <returns>A task that represents the asynchronous test operation.</returns>
     [TestMethod]
-    [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite ExecuteScalarInterpolated")]
-    [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite ExecuteScalarInterpolatedAsync")]
-    [DataRow(DatabaseType.SqlServer, false, DisplayName = "SQL Server ExecuteScalarInterpolated")]
-    [DataRow(DatabaseType.SqlServer, true, DisplayName = "SQL Server ExecuteScalarInterpolatedAsync")]
+    [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite ExecuteScalarInterpolated.")]
+    [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite ExecuteScalarInterpolatedAsync.")]
+    [DataRow(DatabaseType.SqlServer, false, DisplayName = "SQL Server ExecuteScalarInterpolated.")]
+    [DataRow(DatabaseType.SqlServer, true, DisplayName = "SQL Server ExecuteScalarInterpolatedAsync.")]
     public async Task TestExecuteScalarInterpolatedAsync(DatabaseType databaseType, bool useAsync)
     {
-        string value = DatabaseUtils.GetMethodName();
-
         using Context context = DatabaseUtils.CreateDatabase(databaseType);
+
+        string value = DatabaseUtils.GetMethodName();
         long id = DatabaseUtils.CreateSingleTestTableEntry(context, value);
+        FormattableString sql = $"SELECT COUNT(*) FROM TestTable1 WHERE ID = {id}";
 
-        string actualValue = useAsync
-            ? await context.Database.ExecuteScalarInterpolatedAsync<string>($"SELECT TestField FROM TestTable1 WHERE ID = {id}").ConfigureAwait(false)
-            : context.Database.ExecuteScalarInterpolated<string>($"SELECT TestField FROM TestTable1 WHERE ID = {id}");
+        int count = useAsync
+            ? await context.Database.ExecuteScalarInterpolatedAsync<int>(sql).ConfigureAwait(false)
+            : context.Database.ExecuteScalarInterpolated<int>(sql);
 
-        Assert.AreEqual(value, actualValue, "Invalid value");
+        Assert.AreEqual(1, count, "Invalid count");
     }
 
     /// <summary>
-    /// Test ExecuteScalarRaw.
+    /// Test ExecuteScalarRaw with params parameter.
     /// </summary>
     /// <param name="databaseType">Database type.</param>
     /// <param name="useAsync">If <see langword="true"/> then tests the async method.</param>
-    /// <returns><see cref="Task"/>.</returns>
+    /// <returns>A task that represents the asynchronous test operation.</returns>
     [TestMethod]
-    [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite ExecuteScalarRaw")]
-    [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite ExecuteScalarRawAsync")]
-    [DataRow(DatabaseType.SqlServer, false, DisplayName = "SQL Server ExecuteScalarRaw")]
-    [DataRow(DatabaseType.SqlServer, true, DisplayName = "SQL Server ExecuteScalarRawAsync")]
-    public async Task TestExecuteScalarRawAsync(DatabaseType databaseType, bool useAsync)
+    [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite ExecuteScalarRaw params.")]
+    [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite ExecuteScalarRawAsync params.")]
+    [DataRow(DatabaseType.SqlServer, false, DisplayName = "SQL Server ExecuteScalarRaw params.")]
+    [DataRow(DatabaseType.SqlServer, true, DisplayName = "SQL Server ExecuteScalarRawAsync params.")]
+    public async Task TestExecuteScalarRawWithParamsAsync(DatabaseType databaseType, bool useAsync)
     {
-        string value = DatabaseUtils.GetMethodName();
-
         using Context context = DatabaseUtils.CreateDatabase(databaseType);
+
+        string value = DatabaseUtils.GetMethodName();
         long id = DatabaseUtils.CreateSingleTestTableEntry(context, value);
+        string sql = "SELECT COUNT(*) FROM TestTable1 WHERE ID = {0}";
 
-        string actualValue = useAsync
-            ? await context.Database.ExecuteScalarRawAsync<string>("SELECT TestField FROM TestTable1 WHERE ID = {0}", parameters: id).ConfigureAwait(false)
-            : context.Database.ExecuteScalarRaw<string>("SELECT TestField FROM TestTable1 WHERE ID = {0}", id);
+        int count = useAsync
+            ? await context.Database.ExecuteScalarRawAsync<int>(sql, parameters: id).ConfigureAwait(false)
+            : context.Database.ExecuteScalarRaw<int>(sql, id);
 
-        Assert.AreEqual(value, actualValue, "Invalid value");
+        Assert.AreEqual(1, count, "Invalid count");
     }
+
+    /// <summary>
+    /// Test ExecuteScalarRaw with IEnumerable parameter.
+    /// </summary>
+    /// <param name="databaseType">Database type.</param>
+    /// <param name="useAsync">If <see langword="true"/> then tests the async method.</param>
+    /// <returns>A task that represents the asynchronous test operation.</returns>
+    [TestMethod]
+    [DataRow(DatabaseType.Sqlite, false, DisplayName = "SQLite ExecuteScalarRaw IEnumerable.")]
+    [DataRow(DatabaseType.Sqlite, true, DisplayName = "SQLite ExecuteScalarRawAsync IEnumerable.")]
+    [DataRow(DatabaseType.SqlServer, false, DisplayName = "SQL Server ExecuteScalarRaw IEnumerable.")]
+    [DataRow(DatabaseType.SqlServer, true, DisplayName = "SQL Server ExecuteScalarRawAsync IEnumerable.")]
+    public async Task TestExecuteScalarRawWithIEnumerableAsync(DatabaseType databaseType, bool useAsync)
+    {
+        using Context context = DatabaseUtils.CreateDatabase(databaseType);
+
+        string value = DatabaseUtils.GetMethodName();
+        long id = DatabaseUtils.CreateSingleTestTableEntry(context, value);
+        List<object> parameters = new () { id };
+        string sql = "SELECT COUNT(*) FROM TestTable1 WHERE ID = {0}";
+
+        int count = useAsync
+            ? await context.Database.ExecuteScalarRawAsync<int>(sql, parameters).ConfigureAwait(false)
+            : context.Database.ExecuteScalarRaw<int>(sql, parameters);
+
+        Assert.AreEqual(1, count, "Invalid count");
+    }
+
 }